pyproject.toml
[project]
name = "ragswift"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "chromadb>=1.5.7",
    "fastapi>=0.136.0",
    "httpx>=0.28.0",
    "langsmith>=0.2.0",
    "llama-index-core>=0.12.0",
    "llama-index-readers-file>=0.4.0",
    "openai>=2.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "ebooklib>=0.18",
    "html2text>=2024.2.26",
    "pypdf>=5.1.0",
    "python-docx>=1.1.0",
    "python-multipart>=0.0.17",
    "sqlmodel>=0.0.22",
    "sse-starlette>=2.1.3",
    "tiktoken>=0.8.0",
    "uvicorn[standard]>=0.35.0",
]
